+++ lib/src/positions/positions.dart
@@ -18,18 +18,28 @@ abstract class Positions {
 class ItemPosition {
   ItemPosition({
     required this.number,
-    required this.position,
-  });
+    @Deprecated('"position" is deprecated and will be removed after v2.0.0, use "x" and "y" instead')
+        double? position,
+    required this.x,
+    required this.y,
+    required this.size,
+  }) : position = position ?? 0;
 
   /// Ordinal number
   final int number;
 
   /// Coordinate
+  @Deprecated(
+      '"position" is deprecated and will be removed after v2.0.0, use "x" and "y" instead')
   final double position;
+
+  final double x;
+  final double y;
+  final double size;
 
   @override
   String toString() {
-    return 'ItemPosition(number: $number, position: $position)';
+    return 'ItemPosition(number: $number, position: $position, x: $x, y: $y, size: $size)';
   }
 
   @override
@@ -38,29 +48,51 @@ class ItemPosition {
 
     return other is ItemPosition &&
         other.number == number &&
-        other.position == position;
+        other.position == position &&
+        other.x == x &&
+        other.y == y &&
+        other.size == size;
   }
 
   @override
-  int get hashCode => number.hashCode ^ position.hashCode;
+  int get hashCode =>
+      number.hashCode ^
+      position.hashCode ^
+      x.hashCode ^
+      y.hashCode ^
+      size.hashCode;
 }
 
 /// InfoItemPosition consists coordinates, order and information about
 /// last (information) item
 class InfoItemPosition extends ItemPosition {
   InfoItemPosition({
-    required int number,
-    required double position,
+    required super.number,
+    @Deprecated('"position" is deprecated and will be removed after v2.0.0, use "x" and "y" instead')
+        double? position,
     required this.amountAdditionalItems,
+    required super.x,
+    required super.y,
+    required super.size,
   })  : assert(amountAdditionalItems != 0),
-        super(number: number, position: position);
+        super(position: position);
+
+  InfoItemPosition.fromItemPosition({
+    required this.amountAdditionalItems,
+    required ItemPosition itemPosition,
+  }) : super(
+          number: itemPosition.number,
+          size: itemPosition.size,
+          x: itemPosition.x,
+          y: itemPosition.y,
+        );
 
   /// Shows amount of additional (hidden) items.
   final int amountAdditionalItems;
 
   @override
   String toString() {
-    return 'InfoItemPosition(number: $number, position: $position, additionalItems: $amountAdditionalItems)';
+    return 'InfoItemPosition(number: $number, position: $position, additionalItems: $amountAdditionalItems, x: $x, y: $y, size: $size)';
   }
 
   @override
@@ -70,12 +102,20 @@ class InfoItemPosition extends ItemPosition {
     return other is InfoItemPosition &&
         other.amountAdditionalItems == amountAdditionalItems &&
         other.number == number &&
-        other.position == position;
+        other.position == position &&
+        other.x == x &&
+        other.y == y &&
+        other.size == size;
   }
 
   @override
   int get hashCode =>
-      number.hashCode ^ position.hashCode ^ amountAdditionalItems.hashCode;
+      number.hashCode ^
+      position.hashCode ^
+      amountAdditionalItems.hashCode ^
+      x.hashCode ^
+      y.hashCode ^
+      size.hashCode;
 }
 
 /// Whether and how to align avatars horizontally.
@@ -111,4 +151,13 @@ class StackLaying {
 
   /// Defines wether an info item is at the top or at the bottom.
   final bool infoItemAtTop;
+}
+
+/// A direction of laying the items
+enum LayoutDirection {
+  /// Items laying in horizontal way
+  horizontal,
+
+  /// Items laying in vertical way
+  vertical,
 }